requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py37h06a4308_0
charset-normalizer=2.0.12=pypi_0
cudatoolkit=11.1.1=h6406543_8
cycler=0.11.0=pypi_0
easydict=1.9=pypi_0
fonttools=4.28.2=pypi_0
idna=3.3=pypi_0
imageio=2.16.1=pypi_0
joblib=1.1.0=pypi_0
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
matplotlib=3.5.0=pypi_0
ncurses=6.3=h7f8727e_2
numpy=1.21.4=pypi_0
opencv-python=4.5.4.60=pypi_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.3.4=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.2=py37h06a4308_0
protobuf=3.19.3=pypi_0
pymesh2=0.3=pypi_0
pyparsing=3.0.6=pypi_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2021.3=pypi_0
pyzmq=22.3.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.27.1=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.2=pypi_0
seaborn=0.11.2=pypi_0
setuptools=58.0.4=py37h06a4308_0
setuptools-scm=6.3.2=pypi_0
six=1.16.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tensorboardx=2.4.1=pypi_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=1.2.2=pypi_0
torch=1.8.0+cu111=pypi_0
torchaudio=0.8.0=pypi_0
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
torchvision=0.9.0+cu111=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
transforms3d=0.3.1=pypi_0
typing-extensions=4.0.0=pypi_0
urllib3=1.26.9=pypi_0
visdom=0.1.8.9=pypi_0
websocket-client=1.3.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
